// models/planet.model.ts
export interface Planet {
    name: string;
    rotationPeriod: string;
    orbitalPeriod: string;
    diameter: string;
    climate: string;
    gravity: string;
    terrain: string;
    surfaceWater: string;
    population: string;
    residents: string[];
    films: string[];
    created: string;
    edited: string;
    url: string;
}

export interface PlanetSpanish {
    nombre: string;
    periodoRotacion: string;
    periodoOrbital: string;
    diametro: string;
    clima: string;
    gravedad: string;
    terreno: string;
    aguaSuperficial: string;
    poblacion: string;
    residentes: string[];
    peliculas: string[];
    creado: string;
    editado: string;
    url: string;
}